+++ src/main/java/lotto/model/lotto/Lotto.java
@@ -3,7 +3,7 @@ package lotto.model.lotto;
 import lotto.error.LottoErrorMessage;
 import lotto.model.Winning;
 import lotto.model.lotto.constant.LottoNumberPrintFormat;
-import lotto.model.lotto_result.DrawNumbers;
+import lotto.model.draw_numbers.DrawNumbers;
 
 import java.util.List;
 import java.util.stream.Collectors;
