+++ src/test/java/lotto/DrawNumberHacker.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import lotto.model.lotto_result.DrawNumbers;
+import lotto.model.draw_numbers.DrawNumbers;
 
 import java.util.List;
