+++ src/test/java/lotto/OutputViewTest.java
@@ -5,8 +5,8 @@ import lotto.model.lotto.Lotto;
 import lotto.model.lotto.constant.LottoNumberPrintFormat;
 import lotto.model.lotto.LottoMachine;
 import lotto.model.lotto.LottoPublisher;
-import lotto.number_generator.NumberGenerator;
-import lotto.number_generator.TestNumberGenerator;
+import lotto.model.number_generator.NumberGenerator;
+import lotto.model.number_generator.TestNumberGenerator;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
