+++ src/test/java/lotto/WinningTest.java
@@ -3,9 +3,9 @@ package lotto;
 import lotto.model.Winning;
 import lotto.model.lotto.Lotto;
 import lotto.model.lotto.LottoPublisher;
-import lotto.model.lotto_result.DrawNumbers;
-import lotto.number_generator.NumberGenerator;
-import lotto.number_generator.TestNumberGenerator;
+import lotto.model.draw_numbers.DrawNumbers;
+import lotto.model.number_generator.NumberGenerator;
+import lotto.model.number_generator.TestNumberGenerator;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
